Add long options to snoing
Implement snoplus/snoing#141

`dest` is redundent with long options defined; `dest` is automatically
set by taking the first long option and replaces `-` with `_`.

This allows for the long option convensions (with dashes) and still have
a valid variable lookup.
    -i or --install-path provides options.install_path

Can give multiple long options, thus:
    -p, --progress, and --update provides options.progress

Following the discussion in the Aug '16 Software Tutorial that
'progress' isn't the *clearest*. This allows for both.