Update ROOT version for RAT
There has been some issues with XCode 8. Known issue with ROOT,
apparently fixed, so updating to latest patch of ROOT 5.

5.34/36 released around April 2016.
XCode 8 released around September 2016.

I believe they more likely have the fix in the latest ROOT 6, as ROOT 5
only has bugfixes now. If the issue was not solved in 5.34/36, it might
be solved in 5.34/38, whenever it will be released.

Here's hoping.